requirements.txt
fastapi==0.120.0
uvicorn[standard]==0.30.0
pydantic==2.10.0
pydantic-settings==2.5.0
PyJWT==2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
africastalking==1.2.9
httpx==0.26.0
requests==2.31.0